requirements.txt
pandas>=2.2.0
plotly>=5.20.0
snowflake-connector-python>=3.12.0
pyspark>=3.5.0
python-dotenv>=1.0.1
